rename EventsCameraGrid and move it to camera folder
