+++ frontend/src/components/camera/CameraGrid.tsx
@@ -3,11 +3,14 @@ import Grow from "@mui/material/Grow";
 import { useTheme } from "@mui/material/styles";
 
 import { CameraCard } from "components/camera/CameraCard";
-import { useCameraStore, useFilteredCameras } from "components/events/utils";
+import {
+  useCameraStore,
+  useFilteredCameras,
+} from "components/camera/useCameraStore";
 import { useCamerasAll } from "lib/api/cameras";
 import * as types from "lib/types";
 
-export function EventsCameraGrid() {
+export function CameraGrid() {
   const theme = useTheme();
   const { toggleCamera } = useCameraStore();
   const camerasAll = useCamerasAll();
